EPSのスコア算出方法変更

+++ backend/domain/repository/stock.go
@@ -2,5 +2,5 @@ package repository
 
 type StockRepository interface {
 	FetchStockInfo(securitiesCode string) (string, string, [][]string, [][]string, [][]string, [][]string, error)
-	CheckEPS(companyPerformances [][]string) *int
+	CheckEPS(companyPerformances [][]string) *float64
 }
