+++ backend/interface/dto/getStockInfo.go
@@ -5,7 +5,7 @@ import (
 	"strings"
 )
 
-func TransferStockInfo(settlementLink string, companyName string, companyPerformances [][]string, financialStatus [][]string, cashFlow [][]string, dividendTrend [][]string, epsScore *int) model.StockInfo {
+func TransferStockInfo(settlementLink string, companyName string, companyPerformances [][]string, financialStatus [][]string, cashFlow [][]string, dividendTrend [][]string, epsScore *float64) model.StockInfo {
 	companyNameTitle := strings.Split(companyName, " ")
 	return model.StockInfo{
 		SettlementLink:      settlementLink,
